Clone repository to temporary directory
When cloning a repository, place it in a temporary directory instead of the home directory. If a temporary directory was created, delete it afterward.

If an explicit/existing directory was cited, do not delete it.

Remove unused function for guessing a ~/workspace/ directory.

Bonus: List remote branches available for checkout.

Fixes: issue #1